[CHORE] Remove 컴포넌트 제거 -> pages에 통합
