--- src/components/RemoveCard/RemoveCard.tsx
@@ -1,49 +1,0 @@
-import * as S from '@/components/RemoveCard/RemoveCardStyle';
-import Button from '@/components/Button/Button';
-import Swal from 'sweetalert2';
-import { useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
-
-function RemoveCard() {
-  const navigate = useNavigate();
-  const handlePage = () => {
-    navigate('workspace');
-  };
-
-  useEffect(() => {
-    const Toast = Swal.mixin({
-      toast: true,
-      position: 'bottom-right',
-      iconColor: 'white',
-      background: 'lightskyblue',
-      color: 'white',
-      customClass: {
-        popup: 'colored-toast',
-      },
-      showConfirmButton: false,
-      timer: 5000,
-      timerProgressBar: true,
-    });
-    Toast.fire({
-      icon: 'info',
-      title: '5초 뒤에 워크스페이스로<br/>이동합니다',
-    });
-  }, []);
-
-  setInterval(() => {
-    location.replace('/workspace/boards');
-  }, 5000);
-
-  return (
-    <S.CardWrapper>
-      <S.GearSvg />
-      <S.CardTitle size="2.5rem">보드를 삭제하였습니다.</S.CardTitle>
-      <S.CardTitle size="2.1rem">삭제된 보드는 30일간 보관되며 이후에 영구삭제됩니다.</S.CardTitle>
-      <Button themes="remove" size="md" border={false} radius="square" onClick={handlePage}>
-        Move to Workspace
-      </Button>
-    </S.CardWrapper>
-  );
-}
-
-export default RemoveCard;
