--- src/pages/Remove/indexStyle.tsx
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  background-color: ${(props) => props.theme.primaryColor_1};
-  width: 100vw;
-  height: 92.5%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
