+++ src/pages/RemoveBoard/indexStyle.tsx
@@ -1,6 +1,15 @@
 import styled from 'styled-components';
 import { ReactComponent as Gear } from '@/assets/images/gear.svg';
 
+export const Container = styled.div`
+  background-color: ${(props) => props.theme.primaryColor_1};
+  width: 100vw;
+  height: 92.5%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
+
 export const CardWrapper = styled.div`
   padding: 30px;
   margin: 0px 30px 70px 30px;
